check if exists

+++ cli/src/test/java/com/liferay/blade/cli/command/InstallExtensionCommandTest.java
@@ -49,7 +49,9 @@ public class InstallExtensionCommandTest {
 
 		File cacheDir = bladeTest.getCacheDir();
 
-		FileUtil.deleteDir(cacheDir.toPath());
+		if (cacheDir.exists()) {
+			FileUtil.deleteDir(cacheDir.toPath());
+		}
 	}
 
 	@Test
